Improve parsing of filter pages
Before it was splitting on the first space, but MediaWiki syntax does not
require a space after the list character. This meant it broke when the line did
not contain a space or contained a space in an unexpected place. It now takes
all of the string after the first character (which we already checked starts
with a list character) and then strips leading/trailing whitespace.